add ServicePeriod.ActiveDates
new method that returns the dates this service period is active.

Clean up docstring of ServicePeriod.GetDateRange

Remove support for passing a date object to IsActiveOn. Too much flexibility is
confusing. It wasn't needed and was the only place datetime.date was used in the
API. It was only useful because it is easier to iterate through date objects
than strings. ActiveDates shows it isn't that hard.

Add helper method for DateStringToDateObject

Review at http://codereview.appspot.com/4241


git-svn-id: http://googletransitdatafeed.googlecode.com/svn/trunk@872 705c8f7f-ad29-0410-9695-93dd3b14b492