stop_sequence in StopTime
Adding stop_sequence to StopTime is the easiest way to differentiate multiple
visits to a stop by a single trip.

This change adds tests for and fixes a few regressions introduced when I moved
the stoptimes to sqlite. The new tests pass with transitfeed.py rev 760 (before
sqlite).

graphserver depends on Stop.trip_index which was removed

Stop.GetStopTimeTrips was broken when a trip visited a stop multiple times

Review at http://codereview.appspot.com/4644


git-svn-id: http://googletransitdatafeed.googlecode.com/svn/trunk@874 705c8f7f-ad29-0410-9695-93dd3b14b492