Loader parameter load_stop_times, cleaned up TestFailureProblemReporter a bit
This is a hack to make load faster when you don't need the times.
I've been using this for awhile and am tired of maintaining the patch ;-)

At the same time I cleaned up TestFailureProblemReporter a bit so it can
eventually replace ExceptionProblemReporterNoExpiration.

Review at http://codereview.appspot.com/7902/show


git-svn-id: http://googletransitdatafeed.googlecode.com/svn/trunk@911 705c8f7f-ad29-0410-9695-93dd3b14b492