Make the usage information more uniform between the various tools.
Factor out the crash handler into util.py and use it in the tools.

Create an OptionParser subclass that lists all usage information when there is
an error and use it in the tools.

In feedvalidator move the parsing of --performance into main

Set the svn:executable prop on kmlwriter.py
http://code.google.com/p/googletransitdatafeed/issues/detail?id=197


Review at http://codereview.appspot.com/154088


git-svn-id: http://googletransitdatafeed.googlecode.com/svn/trunk@1168 705c8f7f-ad29-0410-9695-93dd3b14b492