Bugfix for stop distance validation in trip - float is required
This is a bug fix for the issue http://code.google.com/p/googletransitdatafeed/issues/detail?id=222
The stop lat and lng values are set to None if they are missing and a MissingValue error is reported. The validation code of trips.txt however expected to have a valid stop lat or lng. This caused a crash. Added a check to Trip for not evaluating the trip distance if the stop is invalid.

Review at http://codereview.appspot.com/4526075/

git-svn-id: http://googletransitdatafeed.googlecode.com/svn/trunk@1544 705c8f7f-ad29-0410-9695-93dd3b14b492