Refactor stops to make consolidateStops O(n) not O(n^2)
Instead of doing a n^2 duplicate removal instead use one Map and have it all done
automatically by that.

We use a TreeMap so that the output is sorted by stopId which makes verfication easier
as output should be stable. This might increase the complexity to O(nlgn) but in practice
there didn't appear to be a problem.

Also split Geocoder out into a separate class.

Signed-off-by: Daniel Thomas <drt24@cam.ac.uk>